refactor: split data response safety helpers
