+++ src/adt/http.ts
@@ -27,14 +27,13 @@
  *    No external HTTP dependencies — undici ships with Node.js 22+.
  */
 
-import { Readable } from 'node:stream';
 import type { BTPProxyConfig } from '@arc-mcp/xsuaa-auth/btp';
 import { Agent, Client, type Dispatcher, fetch as undiciFetch } from 'undici';
 import { getCurrentContext } from '../server/context.js';
 import { logger } from '../server/logger.js';
 import { traceHeaders } from '../server/trace-context.js';
+import { capResponseBody, connectivityProxyResponse } from './bounded-response.js';
 import { resolveCookies } from './cookies.js';
-import type { DataResponseBudget } from './data-result-context.js';
 import { resolveAcceptType, resolveContentType } from './discovery.js';
 import { AdtApiError, AdtNetworkError, AdtResponseLimitError } from './errors.js';
 import {
@@ -179,139 +178,6 @@ export interface AdtResponse {
 interface AuthenticationAttemptState {
   rejected: boolean;
   tail: Promise<void>;
-}
-
-function cloneResponseHeaders(headers: Headers): Headers {
-  const copy = new Headers();
-  let legacySetCookie: string | undefined;
-  for (const [key, value] of headers.entries()) {
-    if (key.toLowerCase() === 'set-cookie') {
-      legacySetCookie = value;
-    } else {
-      copy.append(key, value);
-    }
-  }
-  const getSetCookie = (headers as Headers & { getSetCookie?: () => string[] }).getSetCookie;
-  const setCookies = getSetCookie?.call(headers) ?? (legacySetCookie ? [legacySetCookie] : []);
-  for (const cookie of setCookies) copy.append('set-cookie', cookie);
-  return copy;
-}
-
-async function capResponseBody(response: Response, budget: DataResponseBudget): Promise<Response> {
-  const responseBody = response.body;
-  if (responseBody === null) return response;
-  const requestId = getCurrentContext()?.requestId;
-  const contentEncoding = response.headers.get('content-encoding')?.trim().toLowerCase();
-  const contentLength = response.headers.get('content-length')?.trim();
-  if ((!contentEncoding || contentEncoding === 'identity') && contentLength && /^\d+$/.test(contentLength)) {
-    const declaredBytes = Number(contentLength);
-    if (Number.isSafeInteger(declaredBytes)) {
-      try {
-        budget.assertSingleResponseBytes(declaredBytes, requestId);
-      } catch (error) {
-        try {
-          await responseBody.cancel(error);
-        } catch {
-          // Preserve the response-limit error.
-        }
-        throw error;
-      }
-    }
-  }
-  const reader = responseBody.getReader();
-  const reservation = response.ok ? budget.createReservation() : undefined;
-  let seenBytes = 0;
-  let settled = false;
-
-  const release = () => {
-    if (settled) return;
-    settled = true;
-    reservation?.release();
-  };
-
-  const cappedBody = new ReadableStream<Uint8Array>({
-    async pull(controller) {
-      try {
-        const chunk = await reader.read();
-        if (chunk.done) {
-          if (!settled) {
-            settled = true;
-            reservation?.commit();
-          }
-          controller.close();
-          return;
-        }
-        const nextSeenBytes = seenBytes + chunk.value.byteLength;
-        if (reservation) reservation.add(chunk.value.byteLength, requestId);
-        else budget.assertSingleResponseBytes(nextSeenBytes, requestId);
-        seenBytes = nextSeenBytes;
-        controller.enqueue(chunk.value);
-      } catch (error) {
-        release();
-        try {
-          await reader.cancel(error);
-        } catch {
-          // Preserve the original stream/budget error.
-        }
-        controller.error(error);
-      }
-    },
-    async cancel(reason) {
-      release();
-      await reader.cancel(reason);
-    },
-  });
-
-  return new Response(cappedBody, {
-    status: response.status,
-    statusText: response.statusText,
-    headers: cloneResponseHeaders(response.headers),
-  });
-}
-
-function proxyResponseBody(body: Readable, client: Client, signal: AbortSignal): ReadableStream<Uint8Array> {
-  const source = Readable.toWeb(body) as ReadableStream<Uint8Array>;
-  // Keep the same signal that guarded Client.request() active for body transfer.
-  // This also cancels the Node stream through Readable.toWeb() when the MCP call
-  // or the per-fetch deadline expires after response headers have arrived.
-  const reader = source.pipeThrough(new TransformStream<Uint8Array>(), { signal }).getReader();
-  let settled = false;
-
-  const settle = async (error?: Error) => {
-    if (settled) return;
-    settled = true;
-    if (error) await client.destroy(error);
-    else await client.close();
-  };
-
-  return new ReadableStream<Uint8Array>({
-    async pull(controller) {
-      try {
-        const chunk = await reader.read();
-        if (chunk.done) {
-          await settle();
-          controller.close();
-          return;
-        }
-        controller.enqueue(chunk.value);
-      } catch (error) {
-        const cause = error instanceof Error ? error : new Error(String(error));
-        try {
-          await settle(cause);
-        } finally {
-          controller.error(error);
-        }
-      }
-    },
-    async cancel(reason) {
-      const cause = reason instanceof Error ? reason : new Error('The proxy response stream was cancelled.');
-      try {
-        await reader.cancel(reason);
-      } finally {
-        await settle(cause);
-      }
-    },
-  });
 }
 
 /**
@@ -1541,59 +1407,9 @@ export class AdtHttpClient {
         signal,
       });
 
-      const responseHeaders = new Headers();
-      for (const [key, value] of Object.entries(resp.headers)) {
-        if (value !== undefined) {
-          const vals = Array.isArray(value) ? value : [String(value)];
-          for (const v of vals) {
-            responseHeaders.append(key, v);
-          }
-        }
-      }
-
-      // Statuses 204/205/304 are "null body status" per the Fetch spec — the
-      // Response constructor throws if given a non-null body (and `.text()`
-      // yields '' not null). Pass null so a 304 from a conditional GET (ETag
-      // revalidation) survives the proxy path instead of crashing the write
-      // (e.g. edit_method's read-before-write through the Cloud Connector).
-      // (1xx informational statuses are null-body too, but never surface here:
-      // undici's Client doesn't return them as a final status and ADT never emits them.)
       const isNullBodyStatus = resp.statusCode === 204 || resp.statusCode === 205 || resp.statusCode === 304;
-      if (isNullBodyStatus) {
-        resp.body.destroy();
-        return new Response(null, {
-          status: resp.statusCode,
-          headers: responseHeaders,
-        });
-      }
-
-      if (options?.responseBudget) {
-        const contentEncoding = responseHeaders.get('content-encoding')?.trim().toLowerCase();
-        if (contentEncoding && contentEncoding !== 'identity') {
-          const error = new Error(`Unexpected Content-Encoding '${contentEncoding}' on bounded proxy response.`);
-          // We surface the protocol error ourselves. Destroying the unread body with
-          // that error would also emit an unobserved Node stream `error` event.
-          resp.body.destroy();
-          await client.destroy(error);
-          responseOwnsClient = true;
-          throw error;
-        }
-        const response = new Response(proxyResponseBody(resp.body, client, signal), {
-          status: resp.statusCode,
-          headers: responseHeaders,
-        });
-        responseOwnsClient = true;
-        return response;
-      }
-
-      // Keep the established buffered conversion for ordinary proxy requests. Some callers
-      // inspect headers only, so globally transferring client ownership to an unread stream
-      // would leak their short-lived proxy client.
-      const responseBody = await resp.body.text();
-      return new Response(responseBody, {
-        status: resp.statusCode,
-        headers: responseHeaders,
-      });
+      responseOwnsClient = options?.responseBudget !== undefined && !isNullBodyStatus;
+      return await connectivityProxyResponse(resp, client, signal, options?.responseBudget !== undefined);
     } finally {
       if (!responseOwnsClient) await client.close();
     }
